--- src/desktop/assets/onboarding.coffee
@@ -1,2 +1,0 @@
-$ ->
-  require('../apps/personalize/client/onboarding.js').init()
