+++ src/desktop/assets/onboarding.ts
@@ -1,0 +1,1 @@
+import "desktop/apps/personalize/client/onboarding"
